Create assign_contacts_to_vm.py
Create example script demonstrating how to add contact assignments to a virtual machine.

I'm not a python developer, most of this was adapted from the create_vm.py. I couldn't find a lot of documentation on how to create the VM object type, so this is for the next person who comes along with the same problem.